Remove grouped.test
